refactor models and add tests
