--- src/server/controllers/hello.js
@@ -1,3 +1,0 @@
-module.exports = (req, res) => {
-  res.send({ express: 'Hello From Express' });
-};
