+++ src/server/models/postcodes.js
@@ -1,12 +1,20 @@
-const xhr = require("./utils/xhr");
+const xhr = require('./utils/xhr');
+
+// Postcodes :: Contents
+// makePostcodeArray - takes an array of objects from airtable response and returns
+//                     an array of just postcodes e.g. ["E8 3AB", "N4 1TD"]
+// getGeolocation - takes array of postcodes and queries postcodes.io API for geolocation
+//                  returns postcodes.io response (array of objects)
+// makeLatLngArray - takes response from postcodes.io and returns a 2-dimensional array with
+//                   lat and lng for each entry in the response.
 
 const makePostcodeArray = inputArray => inputArray.map(entry => entry.postcode);
 
 const getGeolocation = postcodeArray =>
   new Promise((resolve, reject) => {
     xhr
-      .post("https://api.postcodes.io/postcodes", {
-        postcodes: postcodeArray.slice(0, 100)
+      .post('https://api.postcodes.io/postcodes', {
+        postcodes: postcodeArray.slice(0, 100),
       })
       .then(body => resolve(body.result));
   });
@@ -16,7 +24,7 @@ const makeLatLngArray = inputArray => {
     if (entry.result != null) {
       return [entry.result.latitude, entry.result.longitude];
     } else {
-      return "invalid";
+      return 'invalid';
     }
   });
 };
