+++ src/server/models/updateRecords.test.js
@@ -1,4 +1,4 @@
-const table = require('./table');
+const { updateGeo, joinWithIDs, updateAirtable } = require('./updateRecords');
 const request = require('request');
 
 beforeEach(() => {
@@ -9,41 +9,6 @@ beforeEach(() => {
 afterEach(() => {
   global.console.error.mockRestore();
   global.console.log.mockRestore();
-});
-
-const validOutput = [
-  { id: 'recavzIUC1T3IT1uO', postcode: 'N4 3HF' },
-  { id: 'recqmLe3AjHxnUEPT', postcode: 'N4 3HH' },
-  { id: 'rec6i8mI1ua3ZK6D9', postcode: 'N4 3HQ' },
-];
-
-describe('table :: getNoGeo', () => {
-  it('returns array with just ID and postcode', async () => {
-    const actual = await table.getNoGeo();
-    expect(actual).toEqual(validOutput);
-  });
-});
-
-describe('table :: getAllValidRows', () => {
-  it('returns an array of 9 valid data rows', async () => {
-    const actual = await table.getAllValidRows();
-    expect(actual.length).toBe(9);
-  });
-});
-
-const triggerError = () =>
-  new Promise((resolve, reject) => {
-    table
-      .requestRows('throw_error', (array, record) => {
-        array.push(record.fields);
-      })
-      .then(resolve);
-  });
-
-describe('table :: requestRows', () => {
-  it('handles error from Airtable module', async () => {
-    const actual = await triggerError();
-  });
 });
 
 const airtableResponse = [
@@ -90,34 +55,34 @@ const joinedData = [
 
 describe('table :: joinWithIDs', () => {
   it('returns an array with IDs and geolocation data', async () => {
-    const actual = await table.joinWithIDs(airtableResponse, geolocationOutput);
+    const actual = await joinWithIDs(airtableResponse, geolocationOutput);
     expect(actual).toEqual(joinedData);
   });
 });
 
 describe('table :: updateGeo', () => {
   it('calls API once for each record', async () => {
-    const response = await table.updateGeo(airtableResponse);
+    const response = await updateGeo(airtableResponse);
     expect(response).toBeTruthy();
     expect(request.patch.mock.calls).toHaveLength(4);
     request.patch.mockClear();
   });
   it('returns 0 when called with an empty array', async () => {
-    const response = await table.updateGeo([]);
+    const response = await updateGeo([]);
     expect(response).toBe(0);
   });
 });
 
 describe('table :: updateAirtable', () => {
   it('makes one call to the request module', async () => {
-    const actual = await table.updateAirtable('sdkuhsg', {
+    const actual = await updateAirtable('sdkuhsg', {
       postcode: 'E2 3PD',
     });
     expect(request.patch.mock.calls).toHaveLength(1);
     request.patch.mockClear();
   });
   it('handles error from airtable module', async () => {
-    const actual = await table.updateAirtable('givemeanerror', {
+    const actual = await updateAirtable('givemeanerror', {
       postcode: 'E2 3DD',
     });
     expect(request.patch.mock.calls).toHaveLength(1);
